+++ app/src/main/java/com/drunkenboys/calendarun/util/DatePickerDialog.kt
@@ -1,17 +1,8 @@
-package com.drunkenboys.calendarun
+package com.drunkenboys.calendarun.util
 
-import android.annotation.SuppressLint
 import android.app.DatePickerDialog
 import android.content.Context
-import java.text.SimpleDateFormat
 import java.util.*
-
-@SuppressLint("SimpleDateFormat")
-fun toStringDateFormat(date: Date): String {
-    val dateFormat = SimpleDateFormat("yyyy.MM.dd")
-
-    return dateFormat.format(date)
-}
 
 fun showDatePickerDialog(context: Context, dateSetListener: DatePickerDialog.OnDateSetListener) {
     val cal = Calendar.getInstance()
